Remove unneeded ClasspathScanner, added static registry Create the EDCH key async Since Java classloading can be done in parallel use multiple threads to load all registries to speed up boot time Result: 1600ms new boot time, 3800ms old boot time

+++ gomint-server/src/main/java/io/gomint/server/world/NibbleArray.java
@@ -42,7 +42,7 @@ public class NibbleArray {
      * @param index The index of the nibble to be set
      * @param value The value to set
      */
-    public void set( short index, byte value ) {
+    public void set( int index, byte value ) {
         value &= 0xF;
         this.data[index / 2] &= (byte) ( 0xF << ( ( index + 1 ) % 2 * 4 ) );
         this.data[index / 2] |= (byte) ( value << ( index % 2 * 4 ) );
@@ -54,7 +54,7 @@ public class NibbleArray {
      * @param index The index of the nibble to get
      * @return The nibble's value
      */
-    public byte get( short index ) {
+    public byte get( int index ) {
         return (byte) ( this.data[index / 2] >> ( ( index & 1 ) << 2 ) & 0xF );
     }
